Fix missing hard(er) dependencies on the python interpreter.
Let's not pretend we're compatible with a wide range of python versions
when all we support is the default one (3.9 when building on Debian 11);
those dependencies seem to be derived directly from the C extensions,
and disabling the renaming means we lose this information.

To strengthen dependencies, hardcode the relevant part of dh_python3's
output (without --no-ext-rename) in an extra substitution variable, and
add a check on the default python version so that we're reminded when
the time comes to update those hardcoded dependencies (for Debian 12+).